% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{browse_subject}
\alias{browse_subject}
\title{List subjects under a top-level category}
\usage{
browse_subject(parent_id)
}
\arguments{
\item{parent_id}{Character. A top-level subject ID (e.g. \code{"K3"}).}
}
\value{
A \code{data.frame} with columns: \code{id}, \code{name}, and all variables under
that parent.
}
\description{
List subjects under a top-level category
}
\examples{
\dontrun{
options(bdlR.path = "/Volumes/Samsung T72/Data/API GUS")
browse_subject("K3")
}
}
