% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{get_codebook}
\alias{get_codebook}
\title{Get codebook entries for specific variables}
\usage{
get_codebook(variable_ids)
}
\arguments{
\item{variable_ids}{Integer vector of variable IDs.}
}
\value{
A \code{data.frame} with full metadata for the requested variables.
}
\description{
Get codebook entries for specific variables
}
\examples{
\dontrun{
options(bdlR.path = "/Volumes/Samsung T72/Data/API GUS")
get_codebook(c(6, 7, 8))
}
}
